글 상세보기, 수정, 삭제 버튼은 각각의 feature로 분리
